chg: Rename Consolidated Packages
Consolidated the "path" and "io" packages into fsio. Short for File
System Input Ouput. Should prevent having to giving the package an
alternate name when importing with the Go path package.
