+++ .gitignore
@@ -3,7 +3,7 @@
 /.idea/
 /git/testdata/repo-01/
 /git/tmp/
-/path/tmp/
+/fsio/tmp/
 
 # Files
 # -----
