+++ fsio/io_test.go
@@ -1,4 +1,4 @@
-package io
+package fsio
 
 import (
 	"io"
