+++ fsio/path.go
@@ -1,4 +1,4 @@
-package path
+package fsio
 
 import (
 	"os"
