+++ fsio/testdata/copy-file-01.txt
