+++ fsio/testdata/dir-to-dir-03.bundle
