+++ fsio/testdata/salam.txt
